Add Brighmed, Meddo, and ClearTables
Brighmed is needed for tileratorui, meddo for tilerator, and
ClearTables is just needed on the system for data ingestion.

A cleaner way to install ClearTables would be a .deb, but this
works.

These aren't going to be used right away, but the only impact is
consuming a few MB and it reduces differences while preparing the
schema change

Change-Id: I99168c3914d65bed88ea62caf70e05383ed89d2d